+++ translations/el_GR.ts
@@ -2696,6 +2696,14 @@ This may reduce distractions caused by stopwatch updates.</source>
         <source>Highlight lines containing diagnostics</source>
         <translation type="unfinished"></translation>
     </message>
+    <message>
+        <source>Color of error messages</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Color of warning messages</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>ShortcutItem</name>