+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 -   Use [Kate's Syntax highlighter](https://api.kde.org/frameworks/syntax-highlighting/html/) engine for Code Highlighting. (#1101)
+-   Add options to select error/warning messages colors for message logger. (#521 and #1247)
 
 ### Fixied